fix: export default é uma bosta

+++ src/models/MongoModel.ts
@@ -1,6 +1,6 @@
 import { isValidObjectId, Model, UpdateQuery } from 'mongoose';
 import { ErrorTypes } from '../errors/catalog';
-import IModel from '../interfaces/IModel';
+import { IModel } from '../interfaces/IModel';
 
 abstract class MongoModel<T> implements IModel<T> {
   protected _model:Model<T>;
@@ -34,6 +34,7 @@ abstract class MongoModel<T> implements IModel<T> {
   }
         
   public async delete(_id: string): Promise<T | null> {
+    if (!isValidObjectId(_id)) throw Error(ErrorTypes.InvalidMongoId);
     return this._model.findByIdAndDelete(_id);
   }
 }
